Backup before finalizing.
Jing proposed different ways of calculating the drug pair similiarity:

use quantile normalized similiarity
use the raw rank.

We discussed together after several attempts and decided to stick to my inital plan:

a rank of the drug pair is given by averaging its similarity within two marginal similiarity vector. This is because that rank matrix is not a symmertic matrix.

So rank of a pair of drug A and B is

mean(rank(AB) + rank(BA))